+++ aleksis/core/util/apps.py
@@ -1,5 +1,5 @@
 from importlib import metadata
-from typing import Any, Optional, Sequence
+from typing import TYPE_CHECKING, Any, Optional, Sequence
 
 import django.apps
 from django.contrib.auth.signals import user_logged_in, user_logged_out
@@ -11,6 +11,9 @@ from license_expression import Licensing
 from spdx_license_list import LICENSES
 
 from .core_helpers import copyright_years
+
+if TYPE_CHECKING:
+    from oauth2_provider.models import AbstractApplication
 
 
 class AppConfig(django.apps.AppConfig):
@@ -213,6 +216,30 @@ class AppConfig(django.apps.AppConfig):
         By default, it does nothing.
         """
         pass
+
+    def get_all_scopes(self) -> dict[str, str]:
+        """Return all OAuth scopes and their descriptions for this app."""
+        return {}
+
+    def get_available_scopes(
+        self,
+        application: Optional["AbstractApplication"] = None,
+        request: Optional[HttpRequest] = None,
+        *args,
+        **kwargs,
+    ) -> list[str]:
+        """Return a list of all OAuth scopes available to the request and application."""
+        return list(self.get_all_scopes().keys())
+
+    def get_default_scopes(
+        self,
+        application: Optional["AbstractApplication"] = None,
+        request: Optional[HttpRequest] = None,
+        *args,
+        **kwargs,
+    ) -> list[str]:
+        """Return a list of all OAuth scopes to always include for this request and application."""
+        return []
 
     def _maintain_default_data(self):
         from django.contrib.auth.models import Permission